reinforcement_learning_env/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-reinforcement_learning_env"
version = "0.1.0"
description = "Reinforcement Learning Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    # install from github
    # "openenv-core[core] @ git+https://github.com/meta-pytorch/OpenEnv.git",
    "openenv-core[core]>=0.2.2",
    # Environment-specific dependencies
    "python-chess>=1.999",
    # Agent & Hugging Face dependencies
    "huggingface_hub",
    "transformers",
    "torch",
    "python-dotenv",
    "openai",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m reinforcement_learning_env.server.app
server = "reinforcement_learning_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["reinforcement_learning_env", "reinforcement_learning_env.server"]
package-dir = { "reinforcement_learning_env" = ".", "reinforcement_learning_env.server" = "server" }
